pyproject.toml
[project]
name = "iswii-incluirtitulo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.11",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
    "types-pyjwt>=1.7.1",
    "uvicorn>=0.34.0",
]
